Fixed package names. Added packages. Removed Utilities instance.
