+++ app/src/main/java/rocks/athrow/android_udacity_reviews/Data/FetchTask.java
@@ -1,4 +1,4 @@
-package rocks.athrow.android_udacity_reviews.Data;
+package rocks.athrow.android_udacity_reviews.data;
 
 import android.content.ContentValues;
 import android.content.Context;
@@ -11,8 +11,8 @@ import io.realm.Realm;
 import io.realm.RealmConfiguration;
 import io.realm.RealmQuery;
 import io.realm.RealmResults;
-import rocks.athrow.android_udacity_reviews.ReviewListAdapter;
-import rocks.athrow.android_udacity_reviews.MainActivity;
+import rocks.athrow.android_udacity_reviews.activity.ReviewListAdapter;
+import rocks.athrow.android_udacity_reviews.activity.MainActivity;
 
 /**
  * FetchReviews
