+++ app/src/main/java/rocks/athrow/android_udacity_reviews/Data/RealmFeedback.java
@@ -1,4 +1,4 @@
-package rocks.athrow.android_udacity_reviews.Data;
+package rocks.athrow.android_udacity_reviews.data;
 
 import java.util.Date;
