+++ app/src/main/java/rocks/athrow/android_udacity_reviews/RealmAdapter/RealmModelAdapter.java
@@ -1,4 +1,4 @@
-package rocks.athrow.android_udacity_reviews.RealmAdapter;
+package rocks.athrow.android_udacity_reviews.realmadapter;
 
 import android.content.Context;
 import android.view.View;
