+++ app/src/main/java/rocks/athrow/android_udacity_reviews/RealmAdapter/RealmRecyclerViewAdapter.java
@@ -1,4 +1,4 @@
-package rocks.athrow.android_udacity_reviews.RealmAdapter;
+package rocks.athrow.android_udacity_reviews.realmadapter;
 
 import android.support.v7.widget.RecyclerView;
