+++ app/src/main/java/rocks/athrow/android_udacity_reviews/fragment/DatePickerFragment.java
@@ -1,0 +1,37 @@
+package rocks.athrow.android_udacity_reviews.fragment;
+
+import java.util.Calendar;
+import java.util.Date;
+
+import android.app.DatePickerDialog;
+import android.app.Dialog;
+
+import android.os.Bundle;
+import android.support.annotation.NonNull;
+import android.support.v4.app.DialogFragment;
+
+/**
+ * Created by jose on 7/24/16.
+ */
+public class DatePickerFragment extends DialogFragment {
+    DatePickerDialog.OnDateSetListener ondateSet;
+
+    public DatePickerFragment() {
+    }
+
+    public void setCallBack(DatePickerDialog.OnDateSetListener ondate) {
+        ondateSet = ondate;
+    }
+    @NonNull
+    @Override
+    public Dialog onCreateDialog(Bundle savedInstanceState) {
+        // Use the current date as the default date in the picker
+        final Calendar c = Calendar.getInstance();
+        int year = c.get(java.util.Calendar.YEAR);
+        int month = c.get(Calendar.MONTH);
+        int day = c.get(Calendar.DAY_OF_MONTH);
+        // Create a new instance of DatePickerDialog and return it
+        return new DatePickerDialog(getActivity(), ondateSet, year, month, day);
+    }
+
+}
