+++ app/src/main/res/layout-w900dp/reviews_list.xml
@@ -9,7 +9,7 @@
     android:divider="?android:attr/dividerHorizontal"
     android:orientation="horizontal"
     android:showDividers="middle"
-    tools:context=".MainActivity">
+    tools:context=".activity.MainActivity">
 
     <!--
     This layout is a two-pane layout for the reviews
