+++ app/src/main/res/layout/content_reports.xml
@@ -2,7 +2,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/fragment"
-    android:name="rocks.athrow.android_udacity_reviews.ReportsActivityFragment"
+    android:name="rocks.athrow.android_udacity_reviews.fragment.ReportsActivityFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
