+++ app/src/main/res/layout/fragment_reports.xml
@@ -5,7 +5,7 @@
     android:orientation="vertical"
     android:paddingBottom="@dimen/activity_vertical_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="rocks.athrow.android_udacity_reviews.ReportsActivityFragment"
+    tools:context=".fragment.ReportsActivityFragment"
     tools:showIn="@layout/activity_main">
 
     <TextView
