+++ app/src/main/res/layout/fragment_reviews_detail.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ReviewsDetailActivity">
+    tools:context=".activity.ReviewsDetailActivity">
     <!--============================================================================================
 
      Header
