+++ app/src/main/res/layout/reviews_list.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/swipeContainer"
-    tools:context=".MainActivity"
+    tools:context=".activity.MainActivity"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
