+++ app/src/main/res/menu/menu_reviews_detail.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="rocks.athrow.android_udacity_reviews.ReviewsDetailActivity">
+    tools:context=".activity.ReviewsDetailActivity">
     <item
         android:id="@+id/action_settings"
         android:orderInCategory="100"
